Add files via upload
The challenge was to get a set of data in excel format and clean it for the requirements. The were questions that needed to be answered and the only way to do that was to first clean the data for use. The process involved sorting and filtering the data, removing duplicates, renaming files and replacing spelling mistakes. After that I had to create a pivot tables to answer the questions the owner of the data wanted to know.